+++ packages/gatsby-plugin-orion-edit/src/components/ArticleMetadata.js
@@ -1,45 +1,37 @@
 import React from 'react'
-import Hero from 'gatsby-plugin-orion-view/src/components/Hero/Wrap'
-import { createPropEditor } from 'gatsby-plugin-orion-edit'
+import ArticleMetadata from 'gatsby-plugin-orion-view/src/components/ArticleMetadata/Wrap'
+import createPropEditor from './PropEditor'
 import { makeStyles } from '@material-ui/core'
 
 const useStyles = makeStyles(theme => ({
   root: {
+    maxWidth: 'none',
     border: `1px dashed ${theme.palette.common.black}`,
     margin: '-16px 0 0 -16px',
     padding: '36px 16px 16px',
     width: 'calc(100% + 32px)',
+    height: '100%',
   },
 }))
 
-const HeroEditor = props => {
+const ArticleMetadataEditor = props => {
   const classes = useStyles()
 
   return (
     <div className={classes.root}>
-      <Hero {...props} />
+      <ArticleMetadata {...props} />
     </div>
   )
 }
 
 export default {
-  editor: HeroEditor,
-  preview: Hero,
+  editor: ArticleMetadataEditor,
+  preview: ArticleMetadata,
   settings: createPropEditor({
-    image: {
-      label: 'Image',
+    readTime: {
+      label: 'Read time',
       required: true,
-      type: 'string',
-    },
-    title: {
-      label: 'Title',
-      required: true,
-      type: 'string',
-    },
-    subtitle: {
-      label: 'Subtitle',
-      required: true,
-      type: 'string',
+      type: 'number',
     },
   }),
 }