+++ packages/gatsby-plugin-orion-edit/src/components/EditPage/index.js
@@ -110,7 +110,7 @@ function EditPage({ initialState, onSave }) {
         })),
       })
     }
-  }, [])
+  }, [amArticle, blocks])
 
   const handleSaveDraft = () => {
     savePage(true)