+++ packages/gatsby-plugin-orion-edit/src/selectable-components.js
@@ -1,0 +1,13 @@
+import ArticleContent from './components/ArticleContent'
+import ArticleList from './components/ArticleList'
+import ArticleMetadata from './components/ArticleMetadata'
+import Hero from './components/Hero'
+import ListChildren from './components/ListChildren'
+
+export default {
+  ArticleContent,
+  ArticleList,
+  ArticleMetadata,
+  Hero,
+  ListChildren,
+}